Rename test sample project with spring application properties
