+++ src/test/java/tech/jhipster/lite/module/infrastructure/secondary/FileSystemSpringCommentsCommandsHandlerTest.java
@@ -35,7 +35,7 @@ import tech.jhipster.lite.module.domain.properties.JHipsterProjectFolder;
 class FileSystemSpringCommentsCommandsHandlerTest {
 
   public static final Path EXISTING_SPRING_PROPERTIES = Paths.get(
-    "src/test/resources/projects/project-with-spring-properties/application.properties"
+    "src/test/resources/projects/project-with-spring-application-properties/application.properties"
   );
   private static final FileSystemSpringCommentsCommandsHandler handler = new FileSystemSpringCommentsCommandsHandler();
