+++ src/test/resources/projects/project-with-spring-application-properties/application.properties
